Update README.md
The aim of the project is to sort the list of country. More over we should find the countries that have the longest character  and the one that have the less characters. To be able to build the project we learn about some built in functions like: len() and print(). We also use some list methods like: sort(), sorted(), min(), max(). We also learn about list comprehension. I hope this project can be useful for you.